feat(WalletModal): fix bug on toggle. moved decentralizeMode to walletModal
